Avoid capturing error pointer arguments
In practice it’s the same pointer whether captured or passed into the block. This fixes the use after free warnings, but I’m not sure it wouldn’t be better if they could be fixed another way to get rid of the error pointer parameter of the block and alway capture.